chore: move ContractExecutable type to the root
