+++ soroban-sdk/src/lib.rs
@@ -1254,7 +1254,6 @@ pub mod storage;
 pub mod token;
 mod vec;
 pub use address::{Address, Executable};
-pub use auth::{ContractExecutable, ContractExecutableRef};
 pub use bytes::{Bytes, BytesN};
 pub use map::Map;
 pub use muxed_address::MuxedAddress;
@@ -1268,6 +1267,28 @@ mod tuple;
 
 mod constructor_args;
 pub use constructor_args::ConstructorArgs;
+
+/// Contract executable used for creating a new contract and used in
+/// `CreateContractHostFnContext`.
+#[derive(Clone, Debug)]
+#[contracttype(crate_path = "crate")]
+pub enum ContractExecutable {
+    /// Executable specified by the contract instance as a specific Wasm contract code entry identified by its Wasm sha256 hash.
+    Wasm(BytesN<32>),
+    /// Executable reference via a persistent storage entry owned by this contract or another contract.
+    ExternalRef(ContractExecutableRef),
+}
+
+/// Executable referenced via a persistent storage entry owned by a contract,
+/// either this contract or another contract.
+///
+/// The persistent storage entry owned by the `owner` has the `tag` as its key.
+#[derive(Clone, Debug)]
+#[contracttype(crate_path = "crate")]
+pub struct ContractExecutableRef {
+    pub owner: Address,
+    pub tag: String,
+}
 
 pub mod xdr;
